SecurityConfig, Bcryptencoder and DTOs for remaining models, controllers and services

+++ src/main/java/com/demo/service/CompanyService.java
@@ -3,13 +3,12 @@ package com.demo.service;
 import java.util.List;
 import java.util.Optional;
 
+import com.demo.model.company.CompanyConverter;
+import com.demo.model.company.CompanyDTO;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 
 import com.demo.model.company.Company;
-import com.demo.model.Product.Product;
-import com.demo.model.specifications.CompanySpecification;
 import com.demo.repo.CompanyRepo;
 
 @Service
@@ -30,53 +29,45 @@ public class CompanyService {
 		return companyRepo.findCompanyByType(company_type);
 	}
 	
-	public List<Company> findCompanyByNameOrType(String company, String company_type){
-		List<Company> list = companyRepo.findAll(
-				Specification.where(CompanySpecification.hasCompany(company))
-				.or(CompanySpecification.hasCompanyType(company_type))
-				);
-		return list;
-	}
+//	public List<Company> findCompanyByNameOrType(String company, String company_type){
+//		List<Company> list = companyRepo.findAll(
+//				Specification.where(CompanySpecification.hasCompany(company))
+//				.or(CompanySpecification.hasCompanyType(company_type))
+//				);
+//		return list;
+//	}
 	
-	public String addCompany(Company company) {
-		
-		if(company.getProducts() != null) {
-			for(Product product : company.getProducts()) {
-				product.setCompany(company);
-			}
-		}
-				
+	public String addCompany(CompanyDTO companyDTO) {
+        Company company = CompanyConverter.toCompany(companyDTO);
 		companyRepo.save(company);
 		return "Added Company Successfully";
 	}
 	
-	public void deleteCompany(long id) {
-		companyRepo.deleteById(id);
+	public String deleteCompany(long id) {
+		if(companyRepo.findById(id).isEmpty()){
+            return "Could Not Locate Resource";
+        }
+        companyRepo.deleteById(id);
+        return "Company Removed Successfully";
 	}
 	
-	public String updateProduct(Company company, long id) {
-		Company old = companyRepo.findById(id).orElse(null);
+	public String updateProduct(CompanyDTO companyDTO) {
+		Company old = companyRepo.findById(companyDTO.getCompany_id()).orElse(null);
 		
 		if(old != null) {
 			
-			if(company.getCompany() != null) {
-				old.setCompany(company.getCompany());
+			if(companyDTO.getCompany() != null) {
+				old.setCompany(companyDTO.getCompany());
 			}
 			
-			if(company.getCompany_type() != null) {
-				old.setCompany_type(company.getCompany_type());
+			if(companyDTO.getCompany_type() != null) {
+				old.setCompany_type(companyDTO.getCompany_type());
 			}
-			
-			if(company.getProducts() != null && !company.getProducts().isEmpty()) {
-				for(Product product : company.getProducts()) {
-					old.addProduct(product);
-				}
-			}
-			
-			companyRepo.save(old);
-			return "Updated Company Successfully";
-		}
-		
-		return "Could Not Locate Resource";
-	}
+
+            companyRepo.save(old);
+            return "Updated Company Successfully";
+        }
+        return "Could Not Locate Resource";
+    }
+
 }
